move parser and reader to another packages.
